or1k: Fix multicore stack calculation
Change the type of the stack pointers to enable pointer calculations at byte
granularity, which is needed for the calculation of _or1k_stack_core[c] and
_or1k_exception_stack_core[c] with _or1k_stack_size and
_or1k_exception_stack_size. (util.c:53-54)

+++ libgloss/or1k/or1k-internals.h
@@ -6,17 +6,17 @@
 
 #include "include/or1k-support.h"
 
-extern uint32_t* _or1k_stack_top;
+extern uint8_t* _or1k_stack_top;
 extern size_t _or1k_stack_size;
-extern uint32_t* _or1k_stack_bottom;
+extern uint8_t* _or1k_stack_bottom;
 
-extern uint32_t* _or1k_exception_stack_top;
+extern uint8_t* _or1k_exception_stack_top;
 extern size_t _or1k_exception_stack_size;
-extern uint32_t* _or1k_exception_stack_bottom;
+extern uint8_t* _or1k_exception_stack_bottom;
 
 #ifdef __OR1K_MULTICORE__
-extern uint32_t* *_or1k_stack_core;
-extern uint32_t* *_or1k_exception_stack_core;
+extern uint8_t* *_or1k_stack_core;
+extern uint8_t* *_or1k_exception_stack_core;
 #endif
